bugfix for oversight
Use two separate lists(line 94, 95) for holding keys and values, respectively, instead of a dictionary in order to prevent last item in the loop to override the dictionary value. Now entries with more than one definition should display correctly.
This change also required the removal of newline character in entry (line 122). Otherwise, Stardict editor will complain and will not build the dict files. However, this newline is necessary and it is handled in another loop in line 133.
Added description to *.gls metadata.